refactor: ステージ定義キャッシュを分離
